requirements.txt
runpod
torch==2.1.0+cu118
torchvision==0.16.0+cu118
torchaudio==2.1.0+cu118
numpy==1.24.3
matplotlib==3.7.2
pillow==10.0.0
opencv-python==4.8.1.78
tqdm==4.66.1